Rewind: Support formatted activity log event summaries
The notifications format schema turned out to be practical for reuse
here in the activity log, which also meant that the parser from the
`wrnc` project was available for reuse. This commit pulls it in with
minor modifications.

Additionally in this PR we're overloding the activity log event
`summary` field and the `activityTitle` property to support formatting.
We could very easilly just add another field but for the time being it
seems reasonable to set that field _either_ as a `string` _or_ as an
`object` containing the formatting information; we can adjust this if
it becomes problematic.

In the future the path is to make all `summary` properties the formatted
object, but for now we have to support both in order to make the smooth
transition. Once the API has been changed we can update Calypso to
assume the single data type.

fixes to user/person block

Everything's an object!