SEO Settings: Replace Immutable with native Set
Previously we were using an `Immutable.js` `Set` to store the dirty
fields on the SEO form so we could know which fields to update.

We don't need Immutable though and if we can get rid of our references
to it then we'll be able to drop a large dependency from Calypso.