External media: add transient images to WP library
When working on https://github.com/Automattic/wp-calypso/pull/20008
I found that after inserting a single image and going directly back
to the Pexels external media screen, the transient image added
was still there, making it look like a duplicate image.

This change switches back to the WordPress library screen before
adding the transient image, so that when going back to the
external media source, the transient image isn't there, and
the WP library looks as it should.

This does not change the usage for the user - single images
are still inserted without any further clicks.

Testing:

Make sure you have a connected Google Photos account.
Edit a post,  go to the media library, switch to Google Photos
and add one of your images. The image should automatically
insert into your post. Add more than one image, or something
that is not an image, and you should switch to the WordPress
library screen and stay there once the files are uploaded.